moved files to separate 'api' folder to allow for better grouping
