+++ apps/researcher/src/app/[locale]/(objects)/heritage-object-list.tsx
@@ -1,6 +1,6 @@
 import HeritageObjectCard from './heritage-object-card';
 import {useTranslations} from 'next-intl';
-import {SearchResult} from '@/lib/objects';
+import {SearchResult} from '@/lib/api/objects';
 
 interface Props {
   heritageObjects: SearchResult['heritageObjects'];
