+++ apps/researcher/src/lib/api/objects/definitions.ts
@@ -1,0 +1,20 @@
+import {
+  HeritageObject,
+  SearchResultFilter,
+  SortBy,
+  SortOrder,
+} from '../definitions';
+
+export type SearchResult = {
+  totalCount: number;
+  offset: number;
+  limit: number;
+  sortBy: SortBy;
+  sortOrder: SortOrder;
+  heritageObjects: HeritageObject[];
+  filters: {
+    owners: SearchResultFilter[];
+    types: SearchResultFilter[];
+    subjects: SearchResultFilter[];
+  };
+};
