+++ apps/researcher/src/lib/api/objects/rdf-helpers.test.ts
@@ -1,4 +1,4 @@
-import {ontologyUrl} from './definitions';
+import {ontologyUrl} from '../definitions';
 import {
   getPropertyValue,
   getPropertyValues,
