+++ apps/researcher/src/lib/api/persons/fetcher-result.ts
